% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bed_chrom.R
\name{bed_chrom}
\alias{bed_chrom}
\title{bed_chrom}
\usage{
bed_chrom(bed, seg.size, upp.bound = 200)
}
\arguments{
\item{bed}{BED file with fully processed with 7 columns
("Chromosome","start","end","type","score","strand","signal").}

\item{seg.size}{Size of segments to be used.}
}
\value{
dat chromosome, position and their
estimated chromatin access.
}
\description{
For each chromosome create file that let's the user plot the chromosome
profile for the chromatin access and tumor mutation burden
}
