% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BedMap-package.R
\docType{package}
\name{BedMap-package}
\alias{BedMap}
\alias{BedMap-package}
\title{BedMap: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
