% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrom_profile.R
\name{chrom_profile}
\alias{chrom_profile}
\title{chrom_profile}
\usage{
chrom_profile(bed, gen, seg.size)
}
\arguments{
\item{bed}{BED file with fully processed with 7 columns
("Chromosome","start","end","type","score","strand","signal").}

\item{gen}{complete maf file with chromosome, chromosome start/end, sample ID}

\item{seg.size}{Size of segments to be used.}
}
\value{
dat chromosome, position and their
estimated chromatin access and TMB (scaled).
}
\description{
For each chromosome create file that let's the user plot the chromosome
profile for the chromatin access and tumor mutation burden
}
