% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_plot.R
\name{gene_plot}
\alias{gene_plot}
\title{gene_plot}
\usage{
gene_plot(dat, lim.C, lim.T)
}
\value{
dat a data frame with the name of genes, their positions and their
estimated chromatin access and TMB.
}
\description{
quick plots
}
