% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_chrom_TMB.R
\name{get_gene_chrom_TMB}
\alias{get_gene_chrom_TMB}
\title{get_gene_chrom_TMB}
\usage{
get_gene_chrom_TMB(bed, gen, map, seg.size, cores = 1)
}
\arguments{
\item{bed}{BED file with fully processed with 7 columns
("Chromosome","start","end","type","score","strand","signal").}

\item{gen}{complete maf file with chromosome, chromosome start/end, sample ID}

\item{map}{Start and end positions of all genes in the genome (or interest)}

\item{seg.size}{Size of segments to be used.}
}
\value{
dat a data frame with the name of genes, their positions and their
estimated chromatin access and TMB.
}
\description{
Combines bed files with WGS/WXS/IMPACT genes
for chromatin accessibility and TMB
}
