% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gws_plot.R
\name{gws_plot}
\alias{gws_plot}
\title{gws_plot}
\usage{
gws_plot(dat, lim.C, lim.T)
}
\value{
dat a data frame with the name of genes, their positions and their
estimated chromatin access and TMB.
}
\description{
quick plots
}
