% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cancer.R
\name{process_cancer}
\alias{process_cancer}
\title{process_cancer}
\usage{
process_cancer(cancer, path, output.path)
}
\arguments{
\item{cancer}{name of the cancer to process}

\item{path}{path to the folder containing the data}

\item{output.path}{path to save output}
}
\value{
dat chromosome, position and their
estimated mean chromatin access.
}
\description{
For each chromosome create file that let's the user plot the chromosome
profile for the chromatin access and tumor mutation burden
}
