% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_example.R
\name{run_example}
\alias{run_example}
\title{run_example}
\usage{
run_example(chrom = 2, span = 0.025)
}
\arguments{
\item{chrom}{Numeric integer between 1 and 22 specifying which chromosome to plot}

\item{span}{Numeric value, for the span to be used for smoothing}
}
\value{
Specified chromosome TMB and chromatin access signla overlay
}
\description{
Recreate chromosome 2 plot from 'Cell-of-origin chromatin organization shapes the mutational landscape of cancer'
(https://www.nature.com/articles/nature14221)
}
